protecting api and react routes, making authorized request

+++ frontend/src/components/PostDetails.js
@@ -1,14 +1,25 @@
 import { usePostContext } from "../hooks/usePostContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 //date fns
 import formatDistanceToNow from 'date-fns/formatDistanceToNow'
 
 const PostDetails = ({post}) => {
     const {dispatch} = usePostContext()
+    const {user} = useAuthContext()
 
     const handleClick = async () => {
+        //check if there is user
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/post/' + post._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            //authorization token in headers to be used in middleware
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
